requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==4.0.1
Flask-JWT-Extended==4.6.0
PyMySQL==1.1.1
Werkzeug==3.0.3
python-dotenv==1.0.1
Pillow>=12.0.0
cryptography>=42.0.0
openpyxl>=3.1.0
Flask-Mail